pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "botiquin_windsurf"
version = "0.1.0"
description = "Botiquin Windsurf - A Python project for data processing and analysis"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "pandas>=1.3.0",
    "openpyxl>=3.0.9",
    "unidecode>=1.3.4",
    "langdetect>=1.0.9",
    "rapidfuzz>=2.13.7",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=2.2.2",
    "faiss-cpu>=1.7.4",
    "pyyaml>=6.0",
    "xlsxwriter>=3.0.3"
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=21.12b0",
    "isort>=5.10.1",
    "mypy>=0.931",
    "pylint>=2.12.2"
]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
